--- services/user_management/src/user_management/handler.py
@@ -1,4 +1,0 @@
-from mangum import Mangum
-import app
-
-handler = Mangum(app)
